requirements.txt
streamlit==1.32.0
pandas==2.2.1
gspread==6.0.0
google-auth==2.28.2